+++ eslint.config.js
@@ -1,0 +1,13 @@
+import tseslint from "typescript-eslint";
+
+export default [
+  ...tseslint.configs.recommended,
+  {
+    rules: {
+      "@typescript-eslint/no-non-null-assertion": "off",
+    },
+  },
+  {
+    ignores: ["dist/**", "node_modules/**"],
+  },
+];